added core data and added favirote functionality

+++ TechTest/MyTravelHelper/Modules/SearchTrains/Entity/Stations.swift
@@ -21,7 +21,7 @@ struct Station: Codable {
     var stationLatitude: Double
     var stationLongitude: Double
     var stationCode: String
-    var stationId: Int
+    var stationId: Int16
 
     enum CodingKeys: String, CodingKey {
         case stationDesc = "StationDesc"
